Relocate signup policy
