+++ src/modules/auth/SignUpForm/Policy.tsx
@@ -1,9 +1,14 @@
 import { FormControl, Checkbox, HStack, Text } from '@chakra-ui/react';
 import { FormErrorMessage } from '@src/common/components';
-import { Controller } from 'react-hook-form';
+import { Control, Controller } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
 
-export default function Policy({ name, control }) {
+type Policy = {
+  name: keyof SignUpFormTypes;
+  control: Control<SignUpFormTypes, any>;
+};
+
+export default function Policy({ name, control }: Policy) {
   const { t } = useTranslation('auth');
   return (
     <Controller
