+++ src/modules/auth/index.ts
@@ -1,2 +1,2 @@
 export { default as LoginForm } from './LoginForm/LoginForm';
-export { default as SignUpForm } from './SignUpForm';
+export { default as SignUpForm } from './SignUpForm/SignUpForm';
